requirements.txt
numpy>=1.12.1
chainer>=2.1.0
gym>=0.10.5
chainerrl>=0.4.0
tqdm>=4
PyYAML>=3.12
